requirements.txt
pyrogram==2.0.58
telethon==1.24.0
python-dotenv==1.0.0
loguru==0.5.3